Revert "Z80 bugfixes" (#51)
This reverts commit 75622ef159b8264f9714d9d3fee089a214074080.

+++ src/r6502.cpp
@@ -76,7 +76,7 @@ void r6502::restore(Stream &s)
 #endif
 }
 
-void r6502::raise(int level) {
+void r6502::raise(uint8_t level) {
 	if (level < 0)
 		nmi();
 	else if (!P.bits.I)
